// frontend/src/api/generated/models/NotificationType.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Типы уведомлений
 */
export enum NotificationType {
    PROBLEM_STATUS_CHANGED = 'problem_status_changed',
    PROBLEM_ASSIGNED = 'problem_assigned',
    PROBLEM_COMMENTED = 'problem_commented',
    COMMENT_REPLIED = 'comment_replied',
    PROBLEM_UPVOTED = 'problem_upvoted',
    PROBLEM_VERIFIED = 'problem_verified',
    PROBLEM_REJECTED = 'problem_rejected',
    COMMENT_HIDDEN = 'comment_hidden',
    USER_MENTIONED = 'user_mentioned',
    PROBLEM_SUBSCRIBED = 'problem_subscribed',
}